slightly renamed storage card components
