+++ frontend/src/pages/accounts/details/BaseStorageCard.module.scss
