+++ frontend/src/pages/accounts/details/PublicPrivateStorageCard.module.scss
